Cadeira-de-Visão-Computacional/PBL-2/requirements.txt
streamlit
opencv-python-headless
Pillow
opencv-python
numpy
deepface
matplotlib